+++ src/routes/(app)/[language]/+page.svelte
