+++ src/routes/(app)/[language]/[collection]/+page.svelte
