+++ src/routes/(app)/builder/AddWidget.svelte
