+++ src/routes/(app)/builder/WidgetBuilder.svelte
