+++ src/routes/(app)/builder/WidgetFields.svelte
