+++ src/routes/(app)/collection/Board.svelte
