+++ src/routes/(app)/collection/ModalAddCategory.svelte
