+++ src/routes/(app)/collection/[...collectionName]/+page.svelte
