+++ src/routes/(app)/collection/[...collectionName]/ModalWidgetForm.svelte
