+++ src/routes/(app)/collection/[...collectionName]/edit.svelte
