+++ src/routes/(app)/user/AdminArea.svelte
